+++ Sourcery/CodeGenerated/Typed.generated.swift
@@ -4,12 +4,30 @@
 extension Enum.Case.AssociatedValue {
     var isOptional: Bool { return typeName.isOptional }
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
+    var actualTypeName: TypeName? { return typeName.actualTypeName }
+    var isTuple: Bool { return typeName.isTuple }
 }
 extension Method.Parameter {
     var isOptional: Bool { return typeName.isOptional }
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
+    var actualTypeName: TypeName? { return typeName.actualTypeName }
+    var isTuple: Bool { return typeName.isTuple }
+}
+extension TupleType.Element {
+    var isOptional: Bool { return typeName.isOptional }
+    var unwrappedTypeName: String { return typeName.unwrappedTypeName }
+    var actualTypeName: TypeName? { return typeName.actualTypeName }
+    var isTuple: Bool { return typeName.isTuple }
+}
+extension Typealias {
+    var isOptional: Bool { return typeName.isOptional }
+    var unwrappedTypeName: String { return typeName.unwrappedTypeName }
+    var actualTypeName: TypeName? { return typeName.actualTypeName }
+    var isTuple: Bool { return typeName.isTuple }
 }
 extension Variable {
     var isOptional: Bool { return typeName.isOptional }
     var unwrappedTypeName: String { return typeName.unwrappedTypeName }
+    var actualTypeName: TypeName? { return typeName.actualTypeName }
+    var isTuple: Bool { return typeName.isTuple }
 }