+++ Sourcery/Models/Typealias.swift
@@ -1,14 +1,16 @@
 import Foundation
 
-final class Typealias: NSObject, AutoDiffable {
+final class Typealias: NSObject, AutoDiffable, Typed {
     /// New typealias name
     let aliasName: String
 
     /// Target name
-    let typeName: String
+    let typeName: TypeName
 
-    // sourcery: skipEquality
-    // sourcery: skipDescription
+    // sourcery: skipEquality, skipDescription
+    var type: Type?
+
+    // sourcery: skipEquality, skipDescription
     var parent: Type? {
         didSet {
             parentName = parent?.name
@@ -27,8 +29,9 @@ final class Typealias: NSObject, AutoDiffable {
 
     init(aliasName: String, typeName: String, parent: Type? = nil) {
         self.aliasName = aliasName
-        self.typeName = typeName
+        self.typeName = TypeName(typeName)
         self.parent = parent
         self.parentName = parent?.name
     }
+
 }