+++ CHANGELOG.md
@@ -6,9 +6,15 @@
 ### New Features
 - Variables with default initializer are now supported, e.g. `var variable = Type(...)`
 - Added support for special escaped names in enum cases e.g. `default` or `for`
+- Added support for tuple types and `tuple` filter for variables
+- Enum associated values now have `localName` and `exteranlName` properties.
+- Added `actualTypeName` for `TypeName` that is typealias
+- Added `implements`, `inherits` and `based` filters
 
 ### Bug Fixes
 - Using protocols doesn't expose variables using KVC, which meant some of the typeName properties weren't accessible via Templates, we fixed that using Sourcery itself to generate specific functions.
+- Fixed parsing typealiases for tuples and closure types
+- Fixed parsing associated values of generic types
 
 ### Internal Changes
 - Performed significant refactoring and simplified mutations in parsers